# backend/api-documentation/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Financial Management",
    "description": "Financial Management",
    "version": "1.0.0",
    "contact": {
      "email": "rizalmustaqim150@gmail.com"
    }
  },
  "servers": [
    {
      "url": "localhost:${PORT}"
    }
  ],
  "paths": {
...
    "/pengeluaran": {
      "get": {
        "tags": [
          "Pengeluaran"
        ],
        "summary": "Get All Pengeluaran",
        "operationId": "getAllPengeluaran",
        "parameters": [
          {
            "name": "row",
            "in": "query",
            "description": "Limit Page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 5
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Offset Page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 5
            }
          },
          {
            "name": "kategori_pengeluaran_id",
            "in": "query",
            "description": "Search By Kategori",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 5
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "Search By Id",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 5
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "description": "Search By Start Date",
            "required": false,
            "schema": {
              "type": "string",
              "example": "2024-12-15T20:32:18.430883Z"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "Search By End Date",
            "required": false,
            "schema": {
              "type": "string",
              "example": "2024-12-15T20:32:18.430883Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Authorization": {
                "description": "Bearer token used for authenticated requests",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 2
                      },
                      "jumlah": {
                        "type": "integer",
                        "example": 1
                      },
                      "tanggal": {
                        "type": "string",
                        "example": "2024-12-15T20:32:18.430883Z"
                      },
                      "keterangan": {
                        "type": "string",
                        "example": "Keterangan"
                      },
                      "KategoriPengeluaran": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "example": 1
                            },
                            "nama_kategori": "string",
                            "example": "Kategori"
                          }
                        }
                      }
                    }
                  }
                },
                "example": [
                  {
                    "id": 2,
                    "jumlah": 10000,
                    "tanggal": "2024-12-15T20:32:18.430883Z",
                    "keterangan": "Keterangan",
                    "NamaKategori": [
                      {
                        "id": 1,
                        "nama_kategori": "Kategori"
                      }
                    ]
                  },
                  {
                    "id": 3,
                    "jumlah": 10000,
                    "tanggal": "2024-12-15T20:32:18.430883Z",
                    "keterangan": "Keterangan",
                    "NamaKategori": [
                      {
                        "id": 1,
                        "nama_kategori": "Kategori"
                      }
                    ]
                  }
                ]
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "post": {
        "tags": [
          "Pengeluaran"
        ],
        "summary": "Create Pengeluaran",
        "operationId": "createPengeluaran",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kategori_pengeluaran_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "user_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "jumlah": {
                    "type": "integer",
                    "example": 1000
                  },
                  "tanggal": {
                    "type": "string",
                    "example": "0001-01-01T00:00:00Z"
                  },
                  "keterangan": {
                    "type": "string",
                    "example": "keterangan"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Authorization": {
                "description": "Bearer token used for authenticated requests",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "kategori_pengeluaran_id": {
                    "type": "integer",
                    "example": 1
                    },
                    "user_id": {
                      "type": "integer",
                      "example": 1
                    },
                    "jumlah": {
                      "type": "integer",
                      "example": 1000
                    },
                    "tanggal": {
                      "type": "string",
                      "example": "0001-01-01T00:00:00Z"
                    },
                    "keterangan": {
                      "type": "string",
                      "example": "keterangan"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/pengeluaran/:id": {
      "put": {
        "tags": [
          "Pengeluaran"
        ],
        "summary": "Update Pengeluaran",
        "operationId": "updatePengeluaran",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID Pengeluaran",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kategori_pengeluaran_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "jumlah": {
                    "type": "integer",
                    "example": 30000
                  },
                  "tanggal": {
                    "type": "string",
                    "example": "0001-01-01T00:00:00Z"
                  },
                  "keterangan": {
                    "type": "string",
                    "example": "keterangan"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Authorization": {
                "description": "Bearer token used for authenticated requests",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "kategori_pengeluaran_id": {
                    "type": "integer",
                    "example": 1
                    },
                    "jumlah": {
                      "type": "integer",
                      "example": 30000
                    },
                    "tanggal": {
                      "type": "string",
                      "example": "0001-01-01T00:00:00Z"
                    },
                    "keterangan": {
                      "type": "string",
                      "example": "keterangan"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "delete": {
        "tags": [
          "Pengeluaran"
        ],
        "summary": "Delete Pengeluaran",
        "operationId": "deletePengeluaran",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID Pengeluaran",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Authorization": {
                "description": "Bearer token used for authenticated requests",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "example": null
          },
          "error": {
            "type": "string",
            "example": "General error description"
          },
          "error_messages": {
            "type": "array",
            "example": null
          }
        }
      },
...
    },
    "responses": {
      "NotFound": {
        "description": "The specified resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                }
              ]
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                }
              ]
            }
          }
        }
      },
      "Forbidden": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                }
              ]
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                }
              ]
            }
          }
        }
      },
      "MethodNotAllowed": {
        "description": "MethodNotAllowed",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                }
              ]
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "ServiceUnavailable",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                }
              ]
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                }
              ]
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "UnprocessableEntity",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "example": null
                },
                "error": {
                  "type": "string",
                  "description": "General error message",
                  "example": "Unprocessable entity"
                },
                "error_messages": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "field1 is required",
                    "field2 is invalid format"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}